Implement state update fn and bring back TreeQueue-specific tests
